fix default metadata
